Check for branches before removing repo when cleaning

+++ src/workspace/commands/clean.py
@@ -5,7 +5,7 @@ from time import time
 
 from workspace.commands.helpers import expand_product_groups
 from workspace.config import config
-from workspace.scm import workspace_path, product_name, repos, stat_repo
+from workspace.scm import workspace_path, product_name, repos, stat_repo, all_branches, is_git_repo
 from workspace.utils import silent_run
 
 log = logging.getLogger(__name__)
@@ -49,7 +49,7 @@ def clean(**kwargs):
       modified_time = os.stat(repo).st_mtime
       if name not in keep_products or keep_time and modified_time < keep_time:
         status = stat_repo(repo, return_output=True)
-        if not status or 'nothing to commit' in status and 'working directory clean' in status:
+        if not status or 'nothing to commit' in status and 'working directory clean' in status and is_git_repo(repo) and len(all_branches(repo)) <= 1:
           shutil.rmtree(repo)
           removed_products.append(name)
         else:
